src/GubernareSearch/pyproject.toml
[project]
name = "gubernare-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "selenium>=4.30.0",
    "selenium-stealth>=1.0.6",
    
]
